Update for restructure, use gapic static path methods where sensible
